Refactor rich_status
