+++ cli/src/contract_status/mod.rs
@@ -57,6 +57,7 @@ pub fn get_status_by_args(args: &ArgMatches) -> Result<Option<Status>, Error> {
     } else {
         Ok(Some(status))
     }
+}
 
 #[cfg(test)]
 mod tests {
